+++ tweet.cpp
@@ -3,7 +3,7 @@
 Tweet::Tweet() {
 
 }
-Tweet::Tweet(long id, strand user) {
+Tweet::Tweet(long long id, strand user) {
     m_id = id;
     m_user = user;
 }
